created the OrderItem class and tested it. created also the ProductOrder class, it's empty for now

+++ src/main/java/se/lexicon/huiyi/jpaworkshopordermanagement/entity/Product.java
@@ -6,12 +6,13 @@ import java.util.Objects;
 public class Product {
     private int id;
     private String name;
-    private BigDecimal price;
+    //should be BigDecimal in real life cases
+    private double price;
 
     public Product() {
     }
 
-    public Product(String name, BigDecimal price) {
+    public Product(String name, double price) {
         this.name = name;
         this.price = price;
     }
@@ -29,11 +30,11 @@ public class Product {
         this.name = name;
     }
 
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(BigDecimal price) {
+    public void setPrice(double price) {
         this.price = price;
     }
